src/workers/requirements.txt
yfinance>=0.2.36
supabase>=2.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
